fix : guide 페이지 리다이렉트 로직 분리
