+++ app/guide/page.tsx
@@ -1,0 +1,9 @@
+import GuidePage from '@/pages/guide/guidePage';
+
+export default function Guide() {
+  return (
+    <>
+      <GuidePage></GuidePage>
+    </>
+  );
+}
